migrating plotting code to core package
