+++ fiftyone/core/plots/__init__.py
@@ -1,5 +1,5 @@
 """
-Plotting utilities.
+Plotting framework.
 
 | Copyright 2017-2021, Voxel51, Inc.
 | `voxel51.com <https://voxel51.com/>`_
@@ -14,4 +14,7 @@ from .base import (
     plot_roc_curve,
     location_scatterplot,
     scatterplot,
+    Plot,
+    InteractivePlot,
+    ViewPlot,
 )
